# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export interface LoginResult {
  token: string;
  email: string;
}

export interface SetupStatus {
  needsSetup: boolean;
}

export interface Site {
  id: number;
  slug: string;
  name: string;
  /** @nullable */
  logoUrl: string | null;
  imapHost: string;
  imapEmail: string;
  createdAt: string;
}

export interface SiteInput {
  slug: string;
  name: string;
  /** @nullable */
  logoUrl?: string | null;
  imapHost: string;
  imapEmail: string;
  imapPassword: string;
}

export interface SiteUpdate {
  slug?: string;
  name?: string;
  /** @nullable */
  logoUrl?: string | null;
  imapHost?: string;
  imapEmail?: string;
  /** @nullable */
  imapPassword?: string | null;
}

export interface SitePublicInfo {
  name: string;
  /** @nullable */
  logoUrl: string | null;
  slug: string;
}

export interface NetflixCode {
  id: string;
  profileName: string;
  deviceInfo: string;
  /** @nullable */
  code: string | null;
  receivedAt: string;
  /** @nullable */
  expiresIn: string | null;
}

export interface ConnectionTestResult {
  success: boolean;
  message: string;
}

export interface ApiError {
  error: string;
}