# lib/api-zod/src/generated/types/site.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Site {
  id: number;
  slug: string;
  name: string;
  /** @nullable */
  logoUrl: string | null;
  imapHost: string;
  imapEmail: string;
  createdAt: string;
}