# lib/api-zod/src/generated/types/siteInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SiteInput {
  slug: string;
  name: string;
  /** @nullable */
  logoUrl?: string | null;
  imapHost: string;
  imapEmail: string;
  imapPassword: string;
}