# lib/api-zod/src/generated/types/siteUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SiteUpdate {
  slug?: string;
  name?: string;
  /** @nullable */
  logoUrl?: string | null;
  imapHost?: string;
  imapEmail?: string;
  /** @nullable */
  imapPassword?: string | null;
}